Initial restructure of my profile classes for redux and networking
