+++ js/components/developer/my-profile-screen/editButtonsStyle.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const style = StyleSheet.create({
+  secondaryButtonGroup: {
+    marginRight: 10,
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  buttonFlex: {
+    flex: 1,
+  },
+  logoutButton: {
+    marginTop: 20,
+  },
+});
+
+module.exports = style;
