+++ js/components/developer/payment-info-screen/paymentInfoStyle.js
@@ -1,9 +1,9 @@
 import { StyleSheet } from 'react-native';
 
-const paymentInfoStyle = StyleSheet.create({
+const style = StyleSheet.create({
   addCardButton: {
     padding: 10,
   },
 });
 
-module.exports = paymentInfoStyle;
+module.exports = style;
